mCMS 2.1 (Registration, password encryption, misc extra features)

+++ admin/functions.php
@@ -34,7 +34,7 @@
                     echo "<td> {$cat_id}</td>";
                     echo "<td> {$cat_title}</td>";
                     echo "<td> <a href='categories.php?edit={$cat_id}'>Edit</a></td>";
-                    echo "<td> <a href='categories.php?delete={$cat_id}'>Delete</a></td>";
+                    echo "<td> <a onClick=\"javascript: return confirm('Are you sure? Delete?'); \" href='categories.php?delete={$cat_id}'>Delete</a></td>";
                     echo "</tr>";
               }
             };
