+++ admin/includes/view_all_comments.php
@@ -62,7 +62,7 @@
                     echo "<td>{$comment_date}</td>";
                     echo "<td><a href='comments.php?approve=$comment_id'>Approve</a></td>";
                     echo "<td><a href='comments.php?unapprove=$comment_id'>Unapprove</a></td>";
-                    echo "<td><a href='comments.php?delete=$comment_id'>Delete</a></td>";
+                    echo "<td><a onClick=\"javascript: return confirm('Are you sure? Delete?'); \" href='comments.php?delete=$comment_id'>Delete</a></td>";
                     echo "</tr>";
                 }
                     ?>
