+++ admin/includes/view_all_posts.php
@@ -92,10 +92,11 @@ case 'published':
                     echo "<td>{$post_date}</td>";
                     echo "<td><a href='../post.php?source&p_id={$post_id}' target='_blank'>View Post</a></td>";
                     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                    echo "<td><a onClick=\"javascript: return confirm('Are you sure? Delete?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
                     echo "</tr>";
                 }
                     ?>
+
                     </tbody>
                 </table>
                 </form>
