+++ admin/includes/view_all_users.php
@@ -62,7 +62,7 @@
                     echo "<td><a href='users.php?change_to_admin={$user_id}'>Adminisrator</a></td>";
                     echo "<td><a href='users.php?change_to_sub='$user_id'>Subscriber</a></td>";
                     echo "<td><a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a></td>";
-                    echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+                    echo "<td><a onClick=\"javascript: return confirm('Are you sure? Delete?'); \" href='users.php?delete=$user_id'>Delete</a></td>";
                     echo "</tr>";
                 }
                     ?>
